pyproject.toml
[tool.poetry]
name = "powerpointless"
version = "0.1.0"
description = ""
authors = ["Aatif Syed <aatifsyedyp@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
argcomplete = "^2.0.0"
python-pptx = "^0.6.21"
streamlit = "^1.5.1"
logging-actions = "^0.1.6"
rich = "^11.2.0"

[tool.poetry.dev-dependencies]
pytest = "^6.0"
black = "^21.12b0"
mypy = "^0.931"
pyinstaller = "4.9"

[tool.poetry.scripts]
powerpointless = "powerpointless:_cli_main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
